requirements.txt
# Argos Index - Dependências

# Interface web
streamlit>=1.28.0

# Banco de dados
sqlalchemy>=2.0.0
pymysql>=1.1.0; sys_platform != "win32"

# Monitoramento de arquivos
watchdog>=3.0.0

# Utilitários
python-dateutil>=2.8.2

# Data processing
pandas>=2.0.0

# Environment variables
python-dotenv>=1.0.0

